+++ clang/test/Driver/sycl-offload.cpp
@@ -155,3 +155,8 @@
 // NO_CHECK_SECTION-NOT: clang-offload-bundler{{.*}} "-type=ao" "-targets=sycl-fpga_aocr-intel-unknown"{{.*}} "-check-section"
 // NO_CHECK_SECTION-NOT: clang-offload-bundler{{.*}} "-type=ao" "-targets=sycl-fpga_aocx-intel-unknown"{{.*}} "-check-section"
 // NO_CHECK_SECTION-NOT: clang-offload-bundler{{.*}} "-type=ao" "-targets=sycl-fpga_aocr_emu-intel-unknown"{{.*}} "-check-section"
+
+/// Check -fsycl-targets=spir64 enables addition of -ffine-grained-bitfield-accesses option
+// RUN:   %clangxx -### -fsycl -fsycl-targets=spir64 %s 2>&1 | FileCheck -check-prefixes=CHECK_BITFIELD_OPTION %s
+// CHECK_BITFIELD_OPTION: clang{{.*}} "-ffine-grained-bitfield-accesses"
+
